test(uploads): satisfy repository contracts
