+++ backend/tests/unit/modules/ingestion/artifact_uploads/test_api_chunks.py
@@ -1,3 +1,8 @@
+"""Defend owned chunk writes, assembly, verification, and cleanup.
+
+A failure here means a resumable upload could corrupt or remove staged bytes.
+"""
+
 from __future__ import annotations
 
 import hashlib
@@ -38,104 +43,107 @@ def _receipt(index: int, offset: int, payload: bytes) -> ChunkReceipt:
     )
 
 
-def test_writes_fixed_chunks_idempotently_and_assembles_exact_bytes(
-    tmp_path, monkeypatch
-) -> None:
-    monkeypatch.setattr(
-        "app.modules.ingestion.artifact_uploads.api_chunks.CHUNK_SIZE", 4
+class TestApiChunkUploadAdapter:
+    def test_assembles_idempotent_fixed_chunks(self, tmp_path, monkeypatch) -> None:
+        monkeypatch.setattr(
+            "app.modules.ingestion.artifact_uploads.api_chunks.CHUNK_SIZE", 4
+        )
+        payload = b"abcdefg"
+        session = _session(payload)
+        adapter = ApiChunkUploadAdapter(tmp_path)
+        adapter.session_directory(session.id, create=True)
+
+        adapter.write_chunk(session, _receipt(0, 0, b"abcd"), b"abcd")
+        adapter.write_chunk(session, _receipt(0, 0, b"abcd"), b"abcd")
+        adapter.write_chunk(session, _receipt(1, 4, b"efg"), b"efg")
+        verified = adapter.assemble(session)
+
+        assert verified.materialize().read_bytes() == payload
+        assert verified.size_bytes == len(payload)
+        assert verified.sha256 == hashlib.sha256(payload).hexdigest()
+
+    @pytest.mark.parametrize(
+        ("receipt", "payload", "code"),
+        [
+            (_receipt(-1, 0, b"abcd"), b"abcd", "chunk_index_invalid"),
+            (_receipt(0, 1, b"abcd"), b"abcd", "chunk_offset_invalid"),
+            (_receipt(0, 0, b"abc"), b"abc", "chunk_length_invalid"),
+            (_receipt(0, 0, b"abcd"), b"abce", "chunk_hash_invalid"),
+        ],
     )
-    payload = b"abcdefg"
-    session = _session(payload)
-    adapter = ApiChunkUploadAdapter(tmp_path)
-    adapter.session_directory(session.id, create=True)
+    def test_rejects_invalid_chunk_envelopes(
+        self,
+        tmp_path,
+        monkeypatch,
+        receipt: ChunkReceipt,
+        payload: bytes,
+        code: str,
+    ) -> None:
+        monkeypatch.setattr(
+            "app.modules.ingestion.artifact_uploads.api_chunks.CHUNK_SIZE", 4
+        )
+        session = _session(b"abcdefg")
+        adapter = ApiChunkUploadAdapter(tmp_path)
+        adapter.session_directory(session.id, create=True)
 
-    adapter.write_chunk(session, _receipt(0, 0, b"abcd"), b"abcd")
-    adapter.write_chunk(session, _receipt(0, 0, b"abcd"), b"abcd")
-    adapter.write_chunk(session, _receipt(1, 4, b"efg"), b"efg")
-    verified = adapter.assemble(session)
+        with pytest.raises(ApiChunkError, match=code):
+            adapter.write_chunk(session, receipt, payload)
 
-    assert verified.materialize().read_bytes() == payload
-    assert verified.size_bytes == len(payload)
-    assert verified.sha256 == hashlib.sha256(payload).hexdigest()
+    def test_conflicting_duplicate_preserves_the_first_chunk(
+        self, tmp_path, monkeypatch
+    ) -> None:
+        monkeypatch.setattr(
+            "app.modules.ingestion.artifact_uploads.api_chunks.CHUNK_SIZE", 4
+        )
+        session = _session(b"abcd")
+        adapter = ApiChunkUploadAdapter(tmp_path)
+        directory = adapter.session_directory(session.id, create=True)
+        adapter.write_chunk(session, _receipt(0, 0, b"abcd"), b"abcd")
 
+        with pytest.raises(ApiChunkError, match="chunk_conflict"):
+            adapter.write_chunk(session, _receipt(0, 0, b"abce"), b"abce")
 
-@pytest.mark.parametrize(
-    ("receipt", "payload", "code"),
-    [
-        (_receipt(-1, 0, b"abcd"), b"abcd", "chunk_index_invalid"),
-        (_receipt(0, 1, b"abcd"), b"abcd", "chunk_offset_invalid"),
-        (_receipt(0, 0, b"abc"), b"abc", "chunk_length_invalid"),
-        (_receipt(0, 0, b"abcd"), b"abce", "chunk_hash_invalid"),
-    ],
-)
-def test_rejects_invalid_chunk_envelopes(
-    tmp_path, monkeypatch, receipt: ChunkReceipt, payload: bytes, code: str
-) -> None:
-    monkeypatch.setattr(
-        "app.modules.ingestion.artifact_uploads.api_chunks.CHUNK_SIZE", 4
-    )
-    session = _session(b"abcdefg")
-    adapter = ApiChunkUploadAdapter(tmp_path)
-    adapter.session_directory(session.id, create=True)
+        assert (directory / "00000000.chunk").read_bytes() == b"abcd"
 
-    with pytest.raises(ApiChunkError, match=code):
-        adapter.write_chunk(session, receipt, payload)
+    def test_incomplete_assembly_is_never_published(
+        self, tmp_path, monkeypatch
+    ) -> None:
+        monkeypatch.setattr(
+            "app.modules.ingestion.artifact_uploads.api_chunks.CHUNK_SIZE", 4
+        )
+        session = _session(b"abcdefg")
+        adapter = ApiChunkUploadAdapter(tmp_path)
+        directory = adapter.session_directory(session.id, create=True)
+        adapter.write_chunk(session, _receipt(0, 0, b"abcd"), b"abcd")
 
+        with pytest.raises(ApiChunkError, match="artifact_upload_incomplete"):
+            adapter.assemble(session)
 
-def test_conflicting_duplicate_preserves_the_first_chunk(tmp_path, monkeypatch) -> None:
-    monkeypatch.setattr(
-        "app.modules.ingestion.artifact_uploads.api_chunks.CHUNK_SIZE", 4
-    )
-    session = _session(b"abcd")
-    adapter = ApiChunkUploadAdapter(tmp_path)
-    directory = adapter.session_directory(session.id, create=True)
-    adapter.write_chunk(session, _receipt(0, 0, b"abcd"), b"abcd")
+        assert not (directory / "assembled.upload").exists()
 
-    with pytest.raises(ApiChunkError, match="chunk_conflict"):
-        adapter.write_chunk(session, _receipt(0, 0, b"abce"), b"abce")
+    def test_verified_identity_rejects_replacement(self, tmp_path) -> None:
+        payload = b"verified"
+        session = _session(payload)
+        adapter = ApiChunkUploadAdapter(tmp_path)
+        adapter.session_directory(session.id, create=True)
+        adapter.write_chunk(session, _receipt(0, 0, payload), payload)
+        verified = adapter.assemble(session)
+        verified.path.unlink()
+        verified.path.write_bytes(b"replaced")
 
-    assert (directory / "00000000.chunk").read_bytes() == b"abcd"
+        with pytest.raises(RuntimeError, match="identity_changed"):
+            verified.materialize()
 
+    def test_abort_preserves_unknown_names(self, tmp_path) -> None:
+        payload = b"owned"
+        session = _session(payload)
+        adapter = ApiChunkUploadAdapter(tmp_path)
+        directory = adapter.session_directory(session.id, create=True)
+        adapter.write_chunk(session, _receipt(0, 0, payload), payload)
+        foreign = directory / "not-an-upload-file"
+        foreign.write_bytes(b"preserve")
 
-def test_incomplete_assembly_is_never_published(tmp_path, monkeypatch) -> None:
-    monkeypatch.setattr(
-        "app.modules.ingestion.artifact_uploads.api_chunks.CHUNK_SIZE", 4
-    )
-    session = _session(b"abcdefg")
-    adapter = ApiChunkUploadAdapter(tmp_path)
-    directory = adapter.session_directory(session.id, create=True)
-    adapter.write_chunk(session, _receipt(0, 0, b"abcd"), b"abcd")
+        with pytest.raises(OSError):
+            adapter.abort_owned(session)
 
-    with pytest.raises(ApiChunkError, match="artifact_upload_incomplete"):
-        adapter.assemble(session)
-
-    assert not (directory / "assembled.upload").exists()
-
-
-def test_verified_identity_rejects_replacement(tmp_path) -> None:
-    payload = b"verified"
-    session = _session(payload)
-    adapter = ApiChunkUploadAdapter(tmp_path)
-    adapter.session_directory(session.id, create=True)
-    adapter.write_chunk(session, _receipt(0, 0, payload), payload)
-    verified = adapter.assemble(session)
-    verified.path.unlink()
-    verified.path.write_bytes(b"replaced")
-
-    with pytest.raises(RuntimeError, match="identity_changed"):
-        verified.materialize()
-
-
-def test_abort_removes_only_known_owned_names(tmp_path) -> None:
-    payload = b"owned"
-    session = _session(payload)
-    adapter = ApiChunkUploadAdapter(tmp_path)
-    directory = adapter.session_directory(session.id, create=True)
-    adapter.write_chunk(session, _receipt(0, 0, payload), payload)
-    foreign = directory / "not-an-upload-file"
-    foreign.write_bytes(b"preserve")
-
-    with pytest.raises(OSError):
-        adapter.abort_owned(session)
-
-    assert foreign.read_bytes() == b"preserve"
+        assert foreign.read_bytes() == b"preserve"
